+++ lib/exception.ex
@@ -8,7 +8,7 @@ defmodule DigitalToken.UnknownTokenError do
 
   @impl true
   def exception(id) when is_binary(id) do
-    message = "The token #{inspect id} is not known"
+    message = "The token #{inspect(id)} is not known"
     %__MODULE__{message: message}
   end
 end